Auto-create OctoDNS records for defensive domains
All of our defensively-registered domains should have [the same
config](https://ministryofjustice.github.io/security-guidance/guides/defensive-domain-registration/).

To allow that, we're using Make to auto-generate the config files.
Ideally there'd be a native way to do this in OctoDNS, but there's [an
outstanding issue on their
backlog](https://github.com/github/octodns/issues/7) to allow templating
that's been open for over a year.

As it stands, this requires that the user pass a DEFENSIVE_DOMAINS
environment variable to Make, but I'll fix that in the next commit.

Also add generated files to our gitignore so they don't get committed.