Check if we already own a domain we're registering
If we already own a domain, this script says it's "not available" which,
while true, isn't the most useful or accurate response it could offer.

Instead, we can quickly check to see if it's something we already own
and, if so, skip the other (slower) checks.

I've run all of the domains in `defensive-domains/domains.mk` through
this, and it identified (confirming @AntonyBishop's suggestion) that
we didn't own the following (but could):

    her-majestys-courts-service.uk
    her-majestys-prison-and-probation-service.uk
    legalaidagency.uk
    national-offenders-management-service.uk
    prisonservice.uk
    public-guardian.uk
    publicguardian.uk

And that we don't, and can't, own:

    court.uk
    courts.uk
    courtservice.uk
    hermajestyscourtservice.uk
    hmcs.uk
    hmcts.uk
    laa.uk
    legal-aid.uk
    legalaid.uk
    ministryofjustice.uk
    moj.uk
    noms.uk
    prison.uk
    prisons.uk
    probation.uk
    sentencingcouncil.uk
    tribunal.uk
    tribunals.uk
    yjb.uk
    youth-justice.uk
    youthjustice.uk

In a subsequent commit, I'll remove all of those from
`defensive-domains/domains.mk`.