:recycle: Rename octodns.yaml and config directory
The name of these files aren't intuitive as they could be. Renaming the octodns.yaml file to config and placing at the root gives it a clearer separation between what is octodns config and what is a hostedzone. Renaming the directory config to hostedzones makes sense as it is a directory full of hostedzones.